+++ src/main/java/com/bridgelabz/bookstorebookmodel/model/BookModel.java
@@ -31,8 +31,7 @@ public class BookModel {
 	private String bookName;
 	private String bookAuthor;
 	private String bookDescription;
-//	private MultipartFile bookLogo;
-	private double bookPrice;
+	private long bookPrice;
 	private int bookQuantity;
 	
 	public BookModel(BookDTO bookDTO) {
